Common bridge addresses, common second user, renamed constants.json to config
