+++ e2e-commons/constants.json
@@ -3,8 +3,24 @@
     "address": "0x7FC1442AB55Da569940Eb750AaD2BAA63DA4010E",
     "privateKey": "0x460635eb4ac4287de2d2393985e19b4a9f948ac533453a1044ab8d50330b0df9"
   },
+  "secondUser": {
+    "address": "0xcca2fb44C8C36E51f743269d6F484Fd027B9F9Aa",
+    "privateKey": "0xcf954e07e6a439faf392eb474e95ddb444c2ca444847f2ad6ecc79e1a585e2b8"
+  },
   "validator": {
     "address": "0xaaB52d66283F7A1D5978bcFcB55721ACB467384b",
     "privateKey": "0x8e829f695aed89a154550f30262f1529582cc49dc30eff74a6b491359e0230f9"
+  },
+  "ercToErcBridge": {
+    "home": "0x1feB40aD9420b186F019A717c37f5546165d411E",
+    "foreign": "0x4a58D6d8D416a5fBCAcf3dC52eb8bE8948E25127"
+  },
+  "ercToNativeBidge": {
+    "home": "0x488Af810997eD1730cB3a3918cD83b3216E6eAda",
+    "foreign": "0x488Af810997eD1730cB3a3918cD83b3216E6eAda"
+  },
+  "nativeToErcBridge": {
+    "home": "0x32198D570fffC7033641F8A9094FFDCaAEF42624",
+    "foreign": "0x2B6871b9B02F73fa24F4864322CdC78604207769"
   }
 }
