+++ oracle-e2e/config.json
@@ -1,0 +1,3 @@
+{
+    "contractsPath": "../contracts"
+}
