--- oracle-e2e/constants.json
@@ -1,7 +1,0 @@
-{
-    "temp": {
-        "address": "0xcca2fb44C8C36E51f743269d6F484Fd027B9F9Aa",
-        "privateKey": "0xcf954e07e6a439faf392eb474e95ddb444c2ca444847f2ad6ecc79e1a585e2b8"
-    },
-    "contractsPath": "../contracts"
-}
